Get sensitivities for observables
In the exponential decay test model, the A_tot observable includes both
A_on and A_off, and hence the sum of these is unchanging; this gives
A_tot a sensitivity of 0 to the rate parameter and 1 to the initial
condition A_0. This is a good test for the correct addition of the
sensitivites since the sensitivities for both species add together to
give all 0s or all 1s.

There are a few warts in the implementation, in particular the elegant
logic for making a view on the record array and using the record array
didn't work out. View on the record array, when reshaped, didn't result
in assignment to the correct slices on the underlying record array.

In addition, the multiplication by the sets of coefficients for each
observable is trickier since the sensitivity vector for each individual
parameter either must be calculated separate (as it is here in a for
loop) or by tiling the coefficients to a large array of the appropriate
dimensions for multiplication by all the parameter sensitivity
timecourses at once.